refactor: modify the file position and renaming
